+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This project aims to adhere to [Semantic Versioning](http://semver.org/).
    not exiting early and continuing to close all of the dangling resources.
  - Added check for SHA256 fingerprints. We error and display a message to inform
    the user to use a MD5 fingerprint for the time being.
+ - We now store the multipart upload job id on the metadata of the metadata 
+   object which allows us to get the job id without doing a listing.
 
 ## [2.6.0] - 2016-11-03
 ### Added